Add comments, PointsLoader, cleane code
